Add docs `TextArea::class`.
